requirements.txt
fastapi
websockets
uvicorn
jinja2
bs4
llama-cpp-python
psutil
